1
1

+++ src/main/java/com/example/dto/Activitycate.java
@@ -6,7 +6,7 @@ import lombok.Data;
 public class ActivityCate {
 
      // 클래스 상위 분류 번호 (PK)
-     private long code;
+     private int code;
      // 클래스 상위 분류 명칭(예술, 스포츠, 뷰티, 쿠킹 등)
      private String cate;
